Move useViewControl to hooks folder
Move useViewControl to hooks folder
