+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useKeyboardShortcuts"
 export * from "./useInitialDialog"
 export * from "./useWindowResize"
 export * from "./useLiveStroke"
+export * from "./useViewControl"
